+++ modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/angle_set/cbc_angle_set.cc
@@ -15,7 +15,7 @@ namespace opensn
 {
 
 CBC_AngleSet::CBC_AngleSet(size_t id,
-                           size_t num_groups,
+                           unsigned int num_groups,
                            const SPDS& spds,
                            std::shared_ptr<FLUDS>& fluds,
                            const std::vector<size_t>& angle_indices,
